fix: compute hybrid_loss in FP32 outside autocast to prevent FP16 overflow in Dice loss sum

+++ src/spiking_useg/training/train_loop.py
@@ -69,7 +69,9 @@ def train_one_epoch(
 
             with torch.amp.autocast('cuda'):
                 pred_t = model(x_t)              # (B, 3, H, W) logits
-                loss_t = hybrid_loss(pred_t, y_t)
+                
+            # Compute loss outside autocast in FP32 to prevent FP16 overflow in Dice sum
+            loss_t = hybrid_loss(pred_t.float(), y_t.float())
 
             step_loss = fptt.step(loss_t, scaler=scaler)
             total_loss += step_loss
